Add methods to add generated photos to profile
Okay, I was totally querying wrong, before, which everything was such a pain.

Because of the way data is stored with firebase, if have a persistent ID you might be referring to (like one for Airtable, for example), I now think it's a better idea to have first query to work out the pushkey of the profile you want, then fetch that profile by its push key.

Definitely not the same as using SQL.