Update Github Action (#5)
* Remove node v12 from matrix.

There are some dependencies which currently do not work on versions of node newer than 10.x. We can leave the matrix strategy in place, and as the dependencies get updated, we can revisit the CI pipeline to add newer versions of node back in.

* Update the test step command.

In order to test, we need to run 'npm run test:unit', previously this was missing the 'run'.